+++ src/services/agent/agent.test.ts
@@ -1,7 +1,17 @@
-import { describe, expect, test } from 'vitest'
+import { describe, expect, test, vi, beforeEach } from 'vitest'
 import { AgentService } from './agent'
 import { prisma } from '@/db'
 import { setupTestDbHooks } from '@/db-test-hooks'
+import * as piAgent from '@mariozechner/pi-coding-agent'
+import { type AutofillField } from '@/agent'
+
+vi.mock('@mariozechner/pi-coding-agent', async () => {
+  const actual = await vi.importActual('@mariozechner/pi-coding-agent')
+  return {
+    ...actual,
+    createAgentSession: vi.fn(),
+  }
+})
 
 describe('AgentService', () => {
   setupTestDbHooks()
@@ -217,6 +227,243 @@ describe('AgentService', () => {
       expect(agents[0].user.name).toBe('Bot 1')
       expect(agents[0].provider).toBeDefined()
       expect(agents[0].modelRef).toBeDefined()
+    })
+  })
+
+  describe('agent execution', () => {
+    const mockProviderConfig: PrismaJson.ProviderConfig = {
+      api: 'openai-completions',
+      baseUrl: 'http://localhost:11434/v1',
+      apiKey: 'test-key',
+    }
+
+    const mockModelConfig: PrismaJson.ModelConfig = {
+      reasoning: false,
+      input: ['text'],
+      contextWindow: 1000,
+      maxTokens: 1000,
+      cost: {
+        input: 0,
+        output: 0,
+        cacheRead: 0,
+        cacheWrite: 0,
+      },
+    }
+
+    beforeEach(() => {
+      vi.clearAllMocks()
+    })
+
+    test('autofills with agent', async () => {
+      const db = prisma
+      const svc = new AgentService()
+
+      const team = await db.team.create({
+        data: {
+          name: 'Autofill Team',
+        },
+      })
+
+      const provider = await db.provider.create({
+        data: {
+          name: 'openai',
+          teamId: team.id,
+          config: mockProviderConfig,
+        },
+      })
+
+      const model = await db.model.create({
+        data: {
+          modelId: 'gpt-4',
+          name: 'GPT-4',
+          providerId: provider.id,
+          config: mockModelConfig,
+        },
+      })
+
+      await db.user.create({
+        data: {
+          name: 'Autofiller',
+          email: `autofiller-${Date.now()}@shumai.ai`,
+          type: 'agent',
+          agent: {
+            create: {
+              type: 'autofill',
+              enabled: true,
+              providerId: provider.id,
+              modelId: model.id,
+              config: {
+                provider: 'openai',
+                model: 'gpt-4',
+              },
+            },
+          },
+          teamMembers: {
+            create: {
+              teamId: team.id,
+              role: 'reviewer',
+            },
+          },
+        },
+      })
+
+      const mockSession = {
+        sendUserMessage: vi.fn().mockImplementation(async () => {
+          // eslint-disable-next-line @typescript-eslint/no-explicit-any
+          const tool = (mockSession as any).customTools.find(
+            (t: unknown) => (t as { name: string }).name === 'autofill_metadata',
+          )
+          if (tool) {
+            await tool.execute('call_1', { data: 1 })
+          }
+        }),
+        getSessionStats: vi.fn().mockReturnValue({ tokens: { input: 5, output: 5 } }),
+        getLastAssistantText: vi.fn().mockReturnValue('{"data": 1}'),
+        state: { tools: [], systemPrompt: '', messages: [] },
+        // eslint-disable-next-line @typescript-eslint/no-explicit-any
+        customTools: [] as any[],
+      }
+      // eslint-disable-next-line @typescript-eslint/no-explicit-any
+      ;(piAgent.createAgentSession as any).mockImplementation(async (config: any) => {
+        mockSession.customTools = config.customTools
+        return { session: mockSession }
+      })
+
+      const fields: AutofillField[] = [{ id: 'f1', config: { name: 'Field 1', type: 'text' } }]
+      const resp = await svc.autofill(team.id, 'extract data', [], fields)
+      expect(resp.text).toBe('{"data":1}')
+      expect(resp.usage.inputTokens).toBe(5)
+    })
+
+    test('chats with bot', async () => {
+      const db = prisma
+      const svc = new AgentService()
+
+      const team = await db.team.create({
+        data: {
+          name: 'Chat Bot Team',
+        },
+      })
+
+      const provider = await db.provider.create({
+        data: {
+          name: 'google',
+          teamId: team.id,
+          config: mockProviderConfig,
+        },
+      })
+
+      const model = await db.model.create({
+        data: {
+          modelId: 'gemini',
+          name: 'Gemini',
+          providerId: provider.id,
+          config: mockModelConfig,
+        },
+      })
+
+      await db.user.create({
+        data: {
+          name: 'Chatter',
+          email: `chatter-${Date.now()}@shumai.ai`,
+          type: 'agent',
+          agent: {
+            create: {
+              type: 'chat',
+              enabled: true,
+              providerId: provider.id,
+              modelId: model.id,
+              config: {
+                provider: 'google',
+                model: 'gemini',
+              },
+            },
+          },
+          teamMembers: {
+            create: {
+              teamId: team.id,
+              role: 'reviewer',
+            },
+          },
+        },
+      })
+
+      const mockSession = {
+        sendUserMessage: vi.fn().mockResolvedValue(undefined),
+        getSessionStats: vi.fn().mockReturnValue({ tokens: { input: 5, output: 5 } }),
+        getLastAssistantText: vi.fn().mockReturnValue('mock text'),
+        state: { tools: [], systemPrompt: '', messages: [] },
+      }
+      // eslint-disable-next-line @typescript-eslint/no-explicit-any
+      ;(piAgent.createAgentSession as any).mockResolvedValue({ session: mockSession })
+
+      const resp = await svc.chat(team.id, 'hello')
+      expect(resp.text).toBe('mock text')
+    })
+
+    test('chat with instruction', async () => {
+      const db = prisma
+      const svc = new AgentService()
+
+      const team = await db.team.create({
+        data: {
+          name: 'Pirate Team',
+        },
+      })
+
+      const provider = await db.provider.create({
+        data: {
+          name: 'google',
+          teamId: team.id,
+          config: mockProviderConfig,
+        },
+      })
+
+      const model = await db.model.create({
+        data: {
+          modelId: 'gemini-pro',
+          name: 'Gemini Pro',
+          providerId: provider.id,
+          config: mockModelConfig,
+        },
+      })
+
+      const botUser = await db.user.create({
+        data: { name: 'Pirate', email: 'pirate@shumai.ai', type: 'agent' },
+      })
+
+      const agent = await db.agent.create({
+        data: {
+          id: botUser.id,
+          type: 'chat',
+          enabled: true,
+          providerId: provider.id,
+          modelId: model.id,
+          soul: 'You are a pirate',
+          config: {
+            provider: 'google',
+            model: 'gemini-pro',
+          },
+        },
+      })
+
+      await db.teamMember.create({
+        data: { teamId: team.id, userId: botUser.id, role: 'reviewer' },
+      })
+
+      const mockSession = {
+        sendUserMessage: vi.fn().mockResolvedValue(undefined),
+        getSessionStats: vi.fn().mockReturnValue({ tokens: { input: 10, output: 20 } }),
+        getLastAssistantText: vi.fn().mockReturnValue('Arr matey!'),
+        state: { tools: [], systemPrompt: '', messages: [] },
+      }
+      // eslint-disable-next-line @typescript-eslint/no-explicit-any
+      ;(piAgent.createAgentSession as any).mockResolvedValue({ session: mockSession })
+
+      const resp = await svc.chatWithAgent(team.id, agent.id, 'hello', [], 'Talk like a pirate')
+      expect(resp.text).toBe('Arr matey!')
+      expect(resp.usage.inputTokens).toBe(10)
+      expect(resp.usage.outputTokens).toBe(20)
     })
   })
 })
