+++ src/workflow/local-executor.ts
@@ -3,8 +3,8 @@ import { WorkflowTask, WorkflowTaskStatus, WorkflowTaskType } from '@/generated/
 import { Executor } from './executor'
 import { aiTranscriptionMedia } from './workflows/ai-transcription'
 import { aiEmbeddingMedia } from './workflows/ai-embedding'
-import { aiAutofillMedia } from '@/agent/workflows/ai-autofill'
-import { aiChat } from '@/agent/workflows/ai-chat'
+import { aiAutofillMedia } from './workflows/ai-autofill'
+import { aiChat } from './workflows/ai-chat'
 import { transcodeMedia } from '@/transcode/workflows/transcode'
 import { activities } from './activities/index'
 import { logger } from '@/logger'
